hazeplaya/requirements.txt
flask
flask-socketio
requests
eventlet
yt-dlp
python-dotenv
